refactor: make getColorsForLanguages as a new util outside of controller
