+++ lib/controller.js
@@ -2,14 +2,14 @@
 const _ = require('lodash');
 const axios = require('axios');
 
-const colorList = require('./public/json/colors.json');
-
 // config related variables
 const error_messages = require('./config/error_messages');
 const { statements, mysteryStatements } = require('./config/statements');
 const { GITHUB_API_URL } = require('./config/api');
 
+// load utils
 const { emojiGenerator } = require('./public/js/emoji_generator');
+const { getColorsForLanguages } = require('./public/js/language_colors_generator');
 
 
 const client_id = process.env.CLIENT_ID || false;
@@ -18,15 +18,6 @@ if (!(client_id && client_secret)) {
     throw error_messages.CLIENT_ID_AND_CLIENT_SECRET_MISSING;
 }
 const clientParams = `client_id=${client_id}&client_secret=${client_secret}`;
-
-const getColors = (languages) => {
-    const defaultColor = '#000';
-    const colors = [];
-    Object.keys(languages).forEach((key) => {
-        colors.push(colorList[key] || defaultColor);
-    });
-    return colors;
-};
 
 const reqUser = (username) => {
     const url = `${GITHUB_API_URL}/users/${username}?${clientParams}`;
@@ -136,7 +127,7 @@ exports.index = (req, res) => {
                         type,
                         avatar,
                         languages: JSON.stringify(languages),
-                        colors: JSON.stringify(getColors(languages)),
+                        colors: JSON.stringify(getColorsForLanguages(languages)),
                         showEmoji: emojiGenerator(numberOfRepos),
                         msg,
                         repos,
@@ -160,7 +151,7 @@ exports.index = (req, res) => {
                     type,
                     avatar,
                     languages: JSON.stringify(languages),
-                    colors: JSON.stringify(getColors(languages)),
+                    colors: JSON.stringify(getColorsForLanguages(languages)),
                     msg,
                     repos,
                     statement,
